AB-4 Add posgresql implementation to store addressbook information.

+++ src/main/java/com/cyberexplorer/addressbook/model/ContactAddressRepository.java
@@ -2,6 +2,6 @@ package com.cyberexplorer.addressbook.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactAddressRepository  extends JpaRepository<ContactAddress, Long> {
+public interface ContactAddressRepository  extends JpaRepository<ContactAddress, Integer> {
     ContactAddress findByName(String name);
 }
